@@ -1,8 +1,7 @@
 'use strict';
 
-var React = require('react-native');
-
-var { StyleSheet, View, PropType } = React;
+import React, { Component, PropType } from 'react';
+import { StyleSheet, Text, View } from 'react-native';
 
 var SCOREHEIGHT = 20;
 //var SCORECOLORS = {'-2': '#B0A606', '-1': '#945816', '0': '#D6DCDF', '1': '#A398C5', '2': '#68617F', '3': '#4B4B4B', '4': '#222', '5': '#000'};
