@@ -1,8 +1,8 @@
 'use strict';
 
-var React = require('react-native');
+import React, { Component, PropType } from 'react';
+import { StyleSheet, Text, View } from 'react-native';
 
-var { StyleSheet, View, PropType } = React;
 
 class SquareChart extends View {
 
